\name{landwebNRV-package}
\alias{landwebNRV-package}
\alias{landwebNRV}
\docType{package}
\title{
\packageTitle{landwebNRV}
}
\description{
\packageDescription{landwebNRV}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{landwebNRV}
\packageIndices{landwebNRV}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{landwebNRV}

Maintainer: \packageMaintainer{landwebNRV}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one ~~
~~ per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
